+++ asusctl/examples/anime-gif.rs
@@ -1,6 +1,6 @@
 use std::{env, path::Path, thread::sleep};
 
-use rog_anime::{ActionData, ActionLoader, Sequences};
+use rog_anime::{usb::get_anime_type, ActionData, ActionLoader, Sequences};
 use rog_dbus::RogDbusClientBlocking;
 
 fn main() {
@@ -14,7 +14,8 @@ fn main() {
 
     let path = Path::new(&args[1]);
     let brightness = args[2].parse::<f32>().unwrap();
-    let mut seq = Sequences::new();
+    let anime_type = get_anime_type().unwrap();
+    let mut seq = Sequences::new(anime_type);
     seq.insert(
         0,
         &ActionLoader::AsusAnimation {